+++ CHANGELOG.md
@@ -7,12 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - GPU option for custom VM passthrough (fix for AMD Windows drivers bug).
 - Other distributions support.
+- Ask before overwrite VM/VHD.
 - Git dependency.
 ### Changed
 - Passthrough blueprints default GPU settings.
+- Now script asks before overwriting VM and VHD.
 - Minor semantics change.
 ### Fixed
 - Should work with all GPUs now.
+- No input entries.
 ### Removed
 - MacOS Firmware.
 - Redundant code.